test: ApiTestHelper 생성 #45
