+++ src/test/java/doore/helper/IntegrationTest.java
@@ -1,70 +1,28 @@
 package doore.helper;
 
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.patch;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 import doore.login.utils.GoogleClient;
 import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
 @SpringBootTest
-@AutoConfigureMockMvc
 @Transactional
 @Sql(value = "/clean.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
-public abstract class IntegrationTest {
+public abstract class IntegrationTest extends ApiTestHelper {
+
     @MockBean
     protected GoogleClient googleClient;
-    @Autowired
-    protected MockMvc mockMvc;
-
-    @Autowired
-    protected ObjectMapper objectMapper;
 
     @BeforeEach
     void setUp(final WebApplicationContext context) {
         this.mockMvc = MockMvcBuilders.webAppContextSetup(context)
                 .addFilters(new CharacterEncodingFilter("UTF-8", true))
                 .build();
-    }
-
-    protected ResultActions callPostApi(final String url, final Object value) throws Exception {
-        return mockMvc.perform(post(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(value)));
-    }
-
-    protected ResultActions callDeleteApi(final String url) throws Exception {
-        return mockMvc.perform(delete(url));
-    }
-
-    protected ResultActions callGetApi(final String url) throws Exception {
-        return mockMvc.perform(get(url));
-    }
-
-    protected ResultActions callPutApi(final String url, final Object value) throws Exception {
-        return mockMvc.perform(put(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(value)));
-    }
-    protected ResultActions callPatchApi(final String url, final Object value) throws Exception {
-        return mockMvc.perform(patch(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(value)));
     }
 }
